refactor: move OsSystem into its own file
The System trait and its real implementation shared one file, while
the mock already had its own. Moving OsSystem out leaves system.rs
as the trait alone, which is what callers read to see what System
offers.
