+++ src/platform.rs
@@ -1,7 +1,9 @@
+mod os_system;
 mod stream;
 mod system;
 mod system_mock;
 
+pub use os_system::*;
 pub use stream::*;
 pub use system::*;
 #[cfg(test)]
